selftests/mm: hugetlb_reparenting_test: do not unmount
Patch series "selftests/mm: Improve run_vmtests.sh", v3.

In this series, I'm trying to add 3 missing tests to vm_runtests.sh which
is used to run all the tests in mm suite.  These tests weren't running by
CIs.  While enabling them and through review feedback, I've fixed some
problems in tests as well.  I've found more flakiness in more tests which
I'll be fixing with future patches.

hugetlb-read-hwpoison test is being added where it can only run with newly
added "-d" (destructive) flag only.  Not sure why it is failing again.  So
once it become stable, we can think of moving it to default set of tests
if it doesn't have any side-effect to them.


This patch (of 5):

Do not unmount the cgroup if it wasn't mounted by the test.  The earlier
patch had fixed this for charge_reserved_hugetlb, but not for this test. 
I'm adding fixes tag to that earlier patch.

Link: https://lkml.kernel.org/r/20240125154608.720072-1-usama.anjum@collabora.com
Link: https://lkml.kernel.org/r/20240125154608.720072-2-usama.anjum@collabora.com
Fixes: 209376ed2a84 ("selftests/vm: make charge_reserved_hugetlb.sh work with existing cgroup setting")
Cc: Ryan Roberts <ryan.roberts@arm.com>
Cc: Shuah Khan <shuah@kernel.org>

+++ tools/testing/selftests/mm/hugetlb_reparenting_test.sh
@@ -248,5 +248,7 @@ cleanup
 
 echo ALL PASS
 
-umount $CGROUP_ROOT
-rm -rf $CGROUP_ROOT
+if [[ $do_umount ]]; then
+  umount $CGROUP_ROOT
+  rm -rf $CGROUP_ROOT
+fi
